location service as lambda -> data is now in the body instead of url params
